Migrate api handling to server side
